Code_Frontend/requirements.txt
openpyxl
pandas
streamlit==1.35.0
firebase_admin
asyncio
PyJWT
typing
httpx_oauth
ipython
streamlit_oauth
mysql-connector-python
streamlit-cookies-controller
streamlit_js_eval
streamlit_mic_recorder
streamlit_extras
cloudinary
streamlit-cookies-manager
streamlit_modal
streamlit_navigation_bar